InquiryTable in separate Component file
